RDCC-418 third refactor
